Moved AuthorizationDialog to src/app and files in /components/page to /components/error
