+++ frontend/src/app/AuthorizationDialog.tsx
