+++ frontend/src/app/authorizationConstants.ts
