+++ frontend/src/components/error/FatalErrorPage.tsx
@@ -1,10 +1,11 @@
-import { Error } from "@/components/error/Error";
 import { Footer } from "@/components/footer/Footer";
 import { NavBar } from "@/components/navbar/NavBar";
 import { AppLayout } from "@/components/ui/AppLayout/AppLayout";
 import { t, tx } from "@/lib/i18n";
 import { ErrorReference } from "@/types/generated/openapi";
 import { isDevelopment } from "@/utils/env";
+
+import { Error } from "./Error";
 
 interface FatalErrorPageProps {
   message: string;
